+++ src/main/java/com/devcourse/Main.java
@@ -1,14 +1,15 @@
 package com.devcourse;
 
 import com.devcourse.calc.Calculator;
+import com.devcourse.calc.func.ConverterNoBracket;
 import com.devcourse.calc.repo.CalcHistoryRepository;
-import com.devcourse.calc.func.Converter;
+import com.devcourse.calc.func.ConverterWithBracket;
 import com.devcourse.view.Input;
 import com.devcourse.view.Output;
 
 public class Main {
     public static void main(String[] args) {
-        Calculator calculator = new Calculator(new CalcHistoryRepository(), new Converter());
+        Calculator calculator = new Calculator(new CalcHistoryRepository(), new ConverterNoBracket());
         while (true) {
             try {
                 int menu = selectMenu();
@@ -17,7 +18,6 @@ public class Main {
                 Output.viewResult(e.getMessage());
             }
         }
-
     }
 
     private static int selectMenu() {
