+++ src/main/java/com/devcourse/valid/FormulaValidator.java
@@ -5,7 +5,8 @@ import java.util.regex.Pattern;
 public class FormulaValidator {
 
     public static boolean valid(String formula) {
-        String expressionValidation = "\\(?\\d+(\\s[+\\-*/]\\s\\(?\\d+\\)?)+";
-        return Pattern.matches(expressionValidation, formula);
+        String expressionValidationWithBracket = "\\(?\\d+(\\s[+\\-*/]\\s\\(?\\d+\\)?)+";
+        String expressionValidationNoBracket = "\\d+(\\s[+\\-*/]\\s\\d+)+";
+        return Pattern.matches(expressionValidationNoBracket, formula);
     }
 }
